refactor: move Fullscreen API to com.vaadin.flow.component.fullscreen
Mirrors the Clipboard layout on main: each feature with an action-trigger
API gets its own package with `@NullMarked` package-info, so the public
surface lives behind one import root instead of being mixed in with
`com.vaadin.flow.component.page`.

Moved:
  - Fullscreen, FullscreenBinding, FullscreenState

Stayed:
  - Page.fullscreenSignal() and Page.exitFullscreen() — observation and
    exit don't need a user gesture and remain on Page; Page now imports
    FullscreenState from the new package.

Adds com/vaadin/flow/component/fullscreen/package-info.java with
@NullMarked so the package is non-null by default. Drops the
`onClick_nullComponent_throws` test — NullAway catches null literals on
non-null parameters statically once the package is annotated.

No backwards-compat shims for the old com.vaadin.flow.component.page
location since the Fullscreen API has not been released yet.
