+++ flow-tests/test-root-context/src/main/java/com/vaadin/flow/uitest/ui/TriggerRequestFullscreenView.java
@@ -15,9 +15,9 @@
  */
 package com.vaadin.flow.uitest.ui;
 
+import com.vaadin.flow.component.fullscreen.Fullscreen;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.NativeButton;
-import com.vaadin.flow.component.page.Fullscreen;
 import com.vaadin.flow.router.Route;
 import com.vaadin.flow.uitest.servlet.ViewTestLayout;
